+++ CHANGELOG.md
@@ -9,6 +9,10 @@ API and may have breaking changes during a teeny version change.
 ## [Unreleased]
 ### Added
 - `util::is_proc` & `util::is_method`
+- `rb_enc_compatible` useful for internal string encoding compatibilty checks from
+  which we now have `binding::is_compatible_encoding` and `binding::encoding_compatible`
+- `RString.compatible_with` as the public API for `rb_enc_compatible`
+
 ### Changed
 - Removed Ruby 2.3 support & added 2.6