+++ CHANGELOG.md
@@ -27,6 +27,10 @@ API and may have breaking changes during a teeny version change.
 - Removed Ruby 2.3 support & added 2.6
 - `VM::raise_ex` now accepts `Into<AnyException>` rather than just `AnyException`
 - Refactor internal encoding types
+- Refactor `build.rs` script to use Ruby provided cflags
+
+### Removed
+- pkg-config-rs removed from Rutie and from the build process
 
 ## [0.5.2] - 2018-12-18
 ### Added